requirements.txt
# Advanced Mini Bash - Voice Control Requirements (Phase 3)
# Google Cloud APIs
google-cloud-speech>=2.21.0
google-cloud-translate>=3.11.0
google-cloud-texttospeech>=2.14.0

# Audio processing
pyaudio>=0.2.11
wave>=0.0.2

# Additional utilities
requests>=2.28.0
numpy>=1.21.0

# Optional: For better audio quality
librosa>=0.9.0
soundfile>=0.10.0
